+++ web/src/containers/admin/StateAdmin.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
@@ -6,123 +6,16 @@ import { getStateAffiliations, updateStateAffiliation } from '../../actions/admi
 
 import { getUserStateOrTerritory } from '../../reducers/user.selector';
 
+// Todo: Test these components
+import ManageRoleDialog from './ManageRoleDialog';
+import ConfirmationDialog from './ConfirmationDialog';
+import ManageUserTable from './ManageUserTable';
+
 import { 
   Button, 
-  Dialog,
-  Dropdown,   
   Tabs, 
-  TabPanel, 
-  Table, 
-  TableHead, 
-  TableRow, 
-  TableCell, 
-  TableBody 
+  TabPanel
 } from '@cmsgov/design-system';
- 
-const ManageRoleDialog = ({
-  selectedAffiliation,
-  hideManageModal,
-  handleAffiliationUpdate
-  }) => {
- 
-   // Should these come from the API? 
-  const dropdownOptions = [
-    { label: 'State Coordinator', value: 'State Coordinator' },
-    { label: 'State Contractor', value: 'State Contractor' }
-  ];
-
-  return (
-    <Dialog
-      onExit={hideManageModal}
-      heading="Edit Permissions"
-      actions={[
-        <Button className="ds-u-margin-right--3 ds-u-margin-top--2" onClick={handleAffiliationUpdate}>Save</Button>,
-        <Button variation="danger" onClick={hideManageModal}>Cancel</Button>
-      ]}
-    >
-      <p><strong>Name</strong> {selectedAffiliation.displayName}</p>
-      <p><strong>Phone Number</strong> {selectedAffiliation.primaryPhone}</p>
-      <p><strong>Email</strong> {selectedAffiliation.email}</p>
-      
-      <Dropdown
-        options={dropdownOptions}
-        className=""
-        defaultValue="State Contractor"
-        hint="A State Coordinator is someone who works for a state. A State Contractor is someone who works for a vendor on behalf of the state."
-        size="medium"
-        label="Role"
-        name="medium_dropdown_field"
-      />
-    </Dialog>
-  )
-}
-
-const ConfirmationDialog = ({
-  denyOrRevoke,
-  hideConfirmationModal,
-  handleDenyOrRevoke
-}) => {
-  return (
-    <Dialog
-      onExit={hideConfirmationModal}
-      heading={`Confirm ${denyOrRevoke.charAt(0).toUpperCase() + denyOrRevoke.slice(1)}`}
-      actions={[
-        <Button className="ds-c-button ds-c-button--danger" onClick={handleDenyOrRevoke}>Confirm</Button>,
-        <Button className="ds-c-button ds-c-button--transparent" onClick={hideConfirmationModal}>Cancel</Button>
-      ]}
-    >
-      <p>Are you sure you want to {`${denyOrRevoke}`} access? You will be able to restore access later from the inactive tab.</p>
-    </Dialog>
-  )
-}
-
-// Todo: Need to handle empty state when there is no requests/active/inactive
-const ManageUserTable = ({
-  type,
-  affiliations,
-  isFetching,
-  actions
-}) => {
-  return (
-    <Fragment>
-      {isFetching && (
-        <p>Loading...</p>
-      )}
-      {!isFetching && affiliations.length > 0 && (
-      <Table borderless>
-        <TableHead>
-          <TableRow>
-            <TableCell>Name</TableCell>
-            <TableCell>Email</TableCell>
-            <TableCell>Phone Number</TableCell>
-            {type === 'active' ? <TableCell>Role</TableCell> : null}
-            {type === 'inactive' ? <TableCell>Status</TableCell> : null}
-            <TableCell>Actions</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {affiliations.map((affiliation, index) => (
-            <TableRow key={index}>
-              <TableCell>{affiliation.displayName}</TableCell>
-              <TableCell>{affiliation.email}</TableCell>
-              <TableCell>{affiliation.primaryPhone}</TableCell>
-              {type === 'active' ? <TableCell>{affiliation.role}</TableCell> : null}
-              {type === 'inactive' ? <TableCell>{affiliation.status.charAt(0).toUpperCase() + affiliation.status.slice(1)}</TableCell> : null}
-              <TableCell>
-                <div className="ds-u-display--flex" data-id={affiliation.id}>
-                  {actions && (
-                    actions
-                  )}
-                </div>
-              </TableCell>
-            </TableRow>  
-          ))}
-        </TableBody>
-      </Table>
-      )}
-    </Fragment>
-  )
-}
 
 const StateAdmin = ({ 
   currentState,
@@ -130,18 +23,15 @@ const StateAdmin = ({
   getStateAffiliations: stateAffiliations,
   updateStateAffiliation: updateAffiliation
 }) => {  
-
   const [activeTab, setActiveTab] = useState("pending");
   const [isFetching, setIsFetching] = useState(true);
+  const [isDenied, setIsDenied] = useState(true);
+
+  const [selectedAffiliation, setSelectedAffiliation] = useState();
 
   const [manageModalDisplay, setManageModalDisplay] = useState(false);
   const [confirmationModalDisplay, setConfirmationModalDisplay] = useState(false);
 
-  const [denyOrRevoke, setDenyOrRevoke] = useState("");
-
-  const [selectedAffiliation, setSelectedAffiliation] = useState();
-
-  // Todo: mess around with making the useEffect function async
   useEffect(() => {
     setIsFetching(true);
     async function fetchAffiliations() {
@@ -161,30 +51,27 @@ const StateAdmin = ({
     setSelectedAffiliation(currentAffiliation);
     setManageModalDisplay(true);
   }
-
-  const handleAffiliationUpdate = () => {
-    // Connect this to a new action to post/patch to the API and update the affiliations list
-    async function saveAffiliation() {
-      // How should I get these params?
-      await updateAffiliation(currentState.id, selectedAffiliation.id, 56, "approved");
-    }
-    saveAffiliation().then(() => {
-      // Todo: im forcing a re-fetch of the affiliations, is this OK?
-      stateAffiliations(currentState.id, activeTab);
-      setManageModalDisplay(false); 
-    }).catch(error => {
-      // Todo: update this logging to be better
-      console.log("error", error);
-    })
-  }
-
+  
   const hideManageModal = () => {
     setManageModalDisplay(false);  
   }
 
+  const handleAffiliationUpdate = () => {
+    async function saveAffiliation() {
+      // Todo: Update this to get the appropriate role_id
+      await updateAffiliation(currentState.id, selectedAffiliation.id, 56, "approved");
+    }
+    saveAffiliation().then(() => {
+      stateAffiliations(currentState.id, activeTab);
+      setManageModalDisplay(false); 
+    }).catch(error => {
+      console.log("Error saving the affiliation:", error);
+    })
+  }
+
   const showConfirmationModal = (event) => {
     const denyOrRevoke = event.target.getAttribute("data-deny-or-revoke");
-    setDenyOrRevoke(denyOrRevoke);
+    setIsDenied(denyOrRevoke);
 
     const currentAffiliation = affiliations.find(element => {
       return element.id == event.target.parentNode.getAttribute("data-id")
@@ -193,31 +80,25 @@ const StateAdmin = ({
     setConfirmationModalDisplay(true);
   }
 
-  // refactor this so its not redundant with handleAffiliationUpdate
-  const handleDenyOrRevoke = () => {
-    // Connect this to a new action to post/patch to the API and update the affiliations list
+  const hideConfirmationModal = () => {
+    setConfirmationModalDisplay(false);  
+  }
 
-    // Todo: rethink how to set this deny or revoke
-    // Tif: better to use a bool for isDenied
-    const permissionChangeType = denyOrRevoke === "deny" ? "denied" : "revoked";
+  // Todo: consider refactoring this to reduce redundancy between handleAffiliationUpdate
+  const handleDenyOrRevoke = () => {
+    const permissionChangeType = isDenied ? "denied" : "revoked";
     
-    // How should I get these params?
+    // Todo: Figure out how what role_id should be set to when revoking
     async function saveAffiliation() {
       await updateAffiliation(currentState.id, selectedAffiliation.id, 56, permissionChangeType);
     }
     saveAffiliation().then(() => {
-      // Todo: im forcing a re-fetch of the affiliations, is this OK?
       stateAffiliations(currentState.id, activeTab);
       setConfirmationModalDisplay(false); 
     }).catch(error => {
-      // Todo: update this logging to be better
-      console.log("error", error);
+      console.log("Error saving the affiliation: ", error);
     })
-  }
-  
-  const hideConfirmationModal = () => {
-    setConfirmationModalDisplay(false);  
-  }
+  }  
 
   return (
     <main id="start-main-content" className="ds-l-container ds-u-margin-bottom--5">
